Relocate metrics tracker from src/shared to src/plugin/src/metrics since only plugin code uses it
