+++ src/plugin/tests/metrics-tracker.test.js
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
-import { createMetricsTracker } from "../../shared/metrics-tracker.js";
+import { createMetricsTracker } from "../src/metrics/metrics-tracker.js";
 
 test("snapshot exposes accumulated metrics as plain objects", () => {
     const tracker = createMetricsTracker({ category: "demo" });
